Create a custom ShortID generator and make it available
Create a utilities package and use the initializer to make it available
 to all other scripts by default.
Create a ShortID generator and include it in the utilities package.

The generator is based off of milliseconds since the epoch,
but it will generate unique values by tracking what the last timestamp was.
Warning!  Do not generate more than 1000 ids per second on average,
if you do and the server restarts, you may end up with duplicate ids.
It is filepath friendly, by only containing lowercase letters and numbers.
Every id has 9 digits.